Make some helper APIs public for easier TextTrimming customization
